backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.28.1
pydantic-settings==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
google-cloud-speech==2.33.0
